ai_note_book/backend/requirements.txt
flask
pygments
fastapi~=0.111.0
pydantic~=2.8.2
sqlalchemy~=2.0.31
databases~=0.9.0
aiosqlite
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
uvicorn
python-dotenv~=1.0.1
fastcrud~=0.13.1
openai
requests~=2.32.3
PyYAML~=6.0.1